improve existing boot code TODO: improve architecture

+++ src/include/kernel.h
@@ -34,9 +34,7 @@
 
 #define DEBUG 1
 
-
-void kernel_main(uint64_t magic, uint64_t addr);
-// void debug_init(uint32_t magic, multiboot_info_t * mb);
+void kernel_main(uint32_t magic, uint32_t addr);
 void console_init(void);
 void gdt_init(void);
 void idt_init(void);
